pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Sokolov Fedor <sokol_f4@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
scikit-learn = "1.3.0"
numpy = "1.25.0"
openai = "1.1.0"
pandas = "2.0.3"
loguru = "0.7.0"
aiogram = "2.25.1"
python-dotenv = "1.0.0"
pyyaml = "6.0"
socksio = "1.0.0"
tqdm = "4.65.0"
matplotlib = "3.7.1"
seaborn = "0.12.2"
httpx = "0.25.1"
httpx-socks = "0.8.0"
fire = "0.5.0"
watchdog = "3.0.0"
fuzzywuzzy = "0.18.0"
levenshtein = "0.21.1"
clearml = "1.11.1"
telethon = "1.34.0"
tiktoken = "0.5.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pre-commit = "^3.5.0"
ruff = "^0.1.11"


[tool.ruff]
exclude = [
    "src/app/bot.py", # Because of handlers import
    "src/handlers/commands.py", # Because of unused import (ADMIN_IDS)
    "**/__init__.py",
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    ".venv",
    ".csv"
]

line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
